requirements.txt
appdirs==1.4.4
appnope==0.1.3
audioread==2.1.9
backcall==0.2.0
backports.functools-lru-cache==1.6.4
boto3==1.21.44
botocore==1.24.44
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
celery==5.2.3
# Cython==0.29.28
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.4
Flask==2.0.3
Flask-Cors==3.0.10
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.1
ipykernel==6.13.0
ipython==7.32.0
itsdangerous==2.1.0
jedi==0.18.1
Jinja2==3.0.3
jmespath==1.0.0
joblib==1.1.0
jupyter-client==7.2.2
jupyter-core==4.9.2
librosa==0.9.1
llvmlite==0.38.0
# madmom==0.16.1
MarkupSafe==2.1.0
matplotlib==3.5.2
matplotlib-inline==0.1.3
mido==1.2.10
nest-asyncio==1.5.5
numba==0.55.1
numpy==1.21.5
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pip==21.2.2
pooch==1.6.0
prompt-toolkit==3.0.29
psutil==5.9.0
ptyprocess==0.7.0
pycparser==2.21
pydub==0.25.1
Pygments==2.11.2
pyparsing==3.0.8
pyrubberband==0.3.0
PySoundFile==0.9.0.post1
python-dateutil==2.8.2
pyzmq==22.3.0
redis==4.3.1
requests==2.27.1
resampy==0.2.2
rq==1.10.1
s3transfer==0.5.2
scikit-learn==1.0.2
scipy==1.7.3
setuptools==59.6.0
six==1.16.0
SoundFile==0.10.3.post1
threadpoolctl==3.1.0
tornado==6.1
tqdm==4.63.0
traitlets==5.1.1
typing_extensions==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
Werkzeug==2.0.3
wheel==0.37.1
zipp==3.7.0
